Move LSB to tabber; rename stages to proper number.
