+++ code/projects/app/src/app/stages/e1/e1v2/e1v2.component.html
@@ -1,0 +1,15 @@
+<div class="row">
+  <app-image-upload (imageChange)="onNextImage($event)" />
+  <app-image-download [image$]="newImage$" />
+</div>
+<div class="row">
+  <app-image-display [image$]="newImage$" />
+  <app-image-display [image$]="lsbImage$" />
+</div>
+<textarea
+  rows="5"
+  [value]="textContent"
+  [disabled]="textBound === 0"
+  (input)="onTextChange($event)"
+></textarea>
+{{ textContent.length }}/{{ textBound }}
