+++ code/projects/app/src/app/stages/e2/e2-decode/e2-decode.component.html
