+++ code/projects/app/src/app/stages/e2/e2-encode/e2-encode.component.html
