--- code/projects/app/src/app/stages/e2/e2.component.css
@@ -1,19 +1,0 @@
-:host {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.row {
-  display: flex;
-}
-
-textarea {
-  width: 100%;
-  background: var(--app-white);
-  border: 1px solid var(--app-black);
-}
-
-textarea:disabled {
-  background: var(--app-silver);
-}
