+++ code/projects/app/src/app/stages/e2/e2.component.html
@@ -1,16 +1,9 @@
-<h1>Etap 2: Metoda najmniej znaczącego bitu na tekście v2</h1>
-<div class="row">
-  <app-image-upload (imageChange)="onNextImage($event)" />
-  <app-image-download [image$]="newImage$" />
-</div>
-<div class="row">
-  <app-image-display [image$]="newImage$" />
-  <app-image-display [image$]="lsbImage$" />
-</div>
-<textarea
-  rows="5"
-  [value]="textContent"
-  [disabled]="textBound === 0"
-  (input)="onTextChange($event)"
-></textarea>
-{{ textContent.length }}/{{ textBound }}
+<h1>Etap 2: Dyskretna transformata kosinusowa</h1>
+<app-tabber>
+    <app-tabber-tab name="Kodowanie">
+        <app-e2-encode></app-e2-encode>
+    </app-tabber-tab>
+    <app-tabber-tab name="Dekodowanie">
+        <app-e2-decode></app-e2-decode>
+    </app-tabber-tab>
+</app-tabber>
