+++ code/projects/app/src/app/stages/e2/e2.component.ts
@@ -1,50 +1,18 @@
 import { Component } from "@angular/core";
-import {
-  combineLatest,
-  debounceTime,
-  distinctUntilChanged,
-  map,
-  Subject,
-} from "rxjs";
-import { ImageUploadComponent } from "../../shared/image-upload/image-upload.component";
-import { ImageDisplayComponent } from "../../shared/image-display/image-display.component";
-import { ImageDownloadComponent } from "../../shared/image-download/image-download.component";
-import { lsbHighlight, lsb2extractText, lsb2embedText } from "steg";
+import { TabberComponent, TabberTabComponent } from "../../shared/tabber/tabber.component";
+import { E2EncodeComponent } from "./e2-encode/e2-encode.component";
+import { E2DecodeComponent } from "./e2-decode/e2-decode.component";
 
 @Component({
   selector: "app-e2",
   standalone: true,
   imports: [
-    ImageUploadComponent,
-    ImageDisplayComponent,
-    ImageDownloadComponent,
-  ],
+    TabberComponent,
+    TabberTabComponent,
+    E2EncodeComponent,
+    E2DecodeComponent
+],
   templateUrl: "./e2.component.html",
   styleUrl: "./e2.component.css",
 })
-export class E2Component {
-  protected textContent = "";
-  protected textBound = 0;
-  private readonly textSubject = new Subject<string>();
-  private readonly imageSubject = new Subject<ImageData>();
-  protected readonly newImage$ = combineLatest([
-    this.imageSubject,
-    this.textSubject.pipe(debounceTime(100), distinctUntilChanged()),
-  ]).pipe(map(([image, text]) => lsb2embedText(image, text)));
-  protected readonly lsbImage$ = this.newImage$.pipe(map(lsbHighlight));
-
-  protected onNextImage(image: ImageData): void {
-    this.imageSubject.next(image);
-
-    this.textContent = lsb2extractText(image);
-    this.textSubject.next(this.textContent);
-    this.textBound = Math.floor((image.width * image.height) / 2);
-  }
-
-  protected onTextChange(event: Event): void {
-    const target = event.currentTarget as HTMLTextAreaElement;
-    this.textContent = target.value.slice(0, this.textBound);
-    target.value = this.textContent;
-    this.textSubject.next(this.textContent);
-  }
-}
+export class E2Component {}
