+++ code/projects/app/src/app/stages/e2/helpers/jpeg-decode.ts
