+++ code/projects/app/src/app/stages/e3/slider/slider.component.ts
