+++ code/projects/app/src/app/stages/e4/e4-merge/e4-merge.component.html
