--- code/projects/app/src/app/stages/e5/e5.component.css
@@ -1,5 +1,0 @@
-:host {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
