--- code/projects/app/src/app/stages/e5/e5.component.html
@@ -1,12 +1,0 @@
-<h1>Etap 5: Kryptografia wizualna</h1>
-<app-tabber>
-  <app-tabber-tab name="Konwertuj">
-    <app-e5-convert></app-e5-convert>
-  </app-tabber-tab>
-  <app-tabber-tab name="Rozdziel">
-    <app-e5-split></app-e5-split>
-  </app-tabber-tab>
-  <app-tabber-tab name="Połącz">
-    <app-e5-merge></app-e5-merge>
-  </app-tabber-tab>
-</app-tabber>
